Add trigonometric function, logic gate and Hex, Bin, Oct conversion

+++ app/src/main/java/com/vera/zzl/comp6442_assignment_two_2016/Operation/Addition.java
@@ -19,7 +19,7 @@ public class Addition extends Expressions {
     }
 
     @Override
-    public float evaluate(){
-        return leftPara.evaluate() + rightPara.evaluate();
+    public Float evaluate(){
+        return (float)leftPara.evaluate() + (float)rightPara.evaluate();
     }
 }
