+++ app/src/main/java/com/vera/zzl/comp6442_assignment_two_2016/Operation/Subtraction.java
@@ -19,8 +19,8 @@ public class Subtraction extends Expressions {
     }
 
     @Override
-    public float evaluate(){
-        return leftPara.evaluate() - rightPara.evaluate();
+    public Float evaluate(){
+        return (float)leftPara.evaluate() - (float)rightPara.evaluate();
     }
 
 }
